Fixed imports in testing
