+++ src/ir/mod.rs
@@ -4,9 +4,6 @@ pub mod builder;
 /// Create LLVM module element
 pub mod values;
 
-/// Create operations
-pub mod operations;
-
 /// Create a type
 pub mod types;
