+++ tests/ir_tests/builder_tests.rs
@@ -1,4 +1,6 @@
-use safe_llvm::{analysis::validator::Validator, constants::{DEFAULT_BASIC_BLOCK_NAME, DEFAULT_FUNCTION_NAME}, memory_management::resource_pools::{IRGenerator, TypeTag}, common};
+use safe_llvm::{
+    analysis::validator::Validator, common, constants::{DEFAULT_BASIC_BLOCK_NAME, DEFAULT_FUNCTION_NAME}, ir::core::{IRGenerator, TypeTag}
+};
 
 #[test]
 fn test_builder_creation() {
